+++ src/data-mgt/java/connectors/kcca-source-connector/src/main/java/net/airqo/connectors/KccaSourceConnector.java
@@ -1,6 +1,5 @@
 package net.airqo.connectors;
 
-import net.airqo.connectors.config.KccaSourceConnectorConfig;
 import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.common.config.ConfigException;
 import org.apache.kafka.connect.errors.ConnectException;
@@ -9,15 +8,13 @@ import org.apache.kafka.connect.source.SourceConnector;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import net.airqo.connectors.tasks.KccaSourceTask;
-import net.airqo.connectors.versions.VersionUtil;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static net.airqo.connectors.config.KccaSourceConnectorConfig.CONFIG_DEF;
+import static net.airqo.connectors.KccaSourceConnectorConfig.CONFIG_DEF;
 
 public class KccaSourceConnector extends SourceConnector {
