+++ src/data-mgt/java/connectors/kcca-source-connector/src/main/java/net/airqo/connectors/KccaSourceConnectorConfig.java
@@ -1,4 +1,4 @@
-package net.airqo.connectors.config;
+package net.airqo.connectors;
 
 import org.apache.kafka.common.config.AbstractConfig;
 import org.apache.kafka.common.config.ConfigDef;
@@ -12,10 +12,14 @@ public class KccaSourceConnectorConfig extends AbstractConfig {
     public static final String POLL_INTERVAL = "pollInterval";
     public static final String CLARITY_API_KEY = "clarityApiKey";
     public static final String CLARITY_API_BASE_URL = "clarityApiBaseUrl";
+    public static final String AVERAGE = "average";
+    public static final String TIME_ZONE = "timezone";
 
     public static final ConfigDef CONFIG_DEF = new ConfigDef()
             .define(TOPIC_CONFIG, ConfigDef.Type.STRING, ConfigDef.Importance.HIGH, "The topic to publish data to")
             .define(CLARITY_API_KEY, ConfigDef.Type.STRING, ConfigDef.Importance.HIGH, "Api key")
+            .define(AVERAGE, ConfigDef.Type.STRING, ConfigDef.Importance.HIGH, "Average")
+            .define(TIME_ZONE, ConfigDef.Type.STRING, ConfigDef.Importance.HIGH, "Timezone")
             .define(POLL_INTERVAL, ConfigDef.Type.STRING, ConfigDef.Importance.HIGH, "Pool Interval")
             .define(CLARITY_API_BASE_URL, ConfigDef.Type.STRING, ConfigDef.Importance.HIGH, "Api Base url");
