+++ src/data-mgt/java/connectors/kcca-source-connector/src/main/java/net/airqo/connectors/Utils.java
@@ -1,0 +1,23 @@
+package net.airqo.connectors;
+
+import net.airqo.connectors.models.RawMeasurements;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Utils {
+
+    public static String buildQueryParameters(String average, String timezone){
+
+        return "";
+
+    }
+
+    public static List<RawMeasurements> getMeasurements(String urlString, String apiKey){
+
+        return new ArrayList<>();
+    }
+
+
+}
+
