+++ src/data-mgt/java/connectors/kcca-source-connector/src/main/java/net/airqo/connectors/VersionUtil.java
@@ -1,4 +1,4 @@
-package net.airqo.connectors.versions;
+package net.airqo.connectors;
 
 import java.util.Properties;
